requirements.txt
azure.identity
azure.ai.openai
python-dotenv